requirements.txt
psycopg2==2.7.5
python-dotenv==0.9.1
selenium==3.14.0
SQLAlchemy==1.2.10
urllib3==1.23
